requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
faiss-cpu==1.7.4
sentence-transformers==2.2.2
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.10.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
pypandoc==1.12
beautifulsoup4==4.12.2
lxml==4.9.3

# Web interface
streamlit==1.29.0
streamlit-chat==0.1.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
chromadb==0.4.18
requests==2.31.0

# Optional: For better performance
torch==2.1.0
transformers==4.35.2

